Version 0.5.1

+++ cloudwok/admin.php
@@ -1,7 +1,40 @@
 <?php
     define( 'BLOCK_LOAD', true );
-    require_once( $_SERVER['DOCUMENT_ROOT'] . '/wp-config.php' );
-    require_once( $_SERVER['DOCUMENT_ROOT'] . '/wp-includes/wp-db.php' );
+    define('MYPATH', dirname(__FILE__));
+    if ( file_exists( $_SERVER['DOCUMENT_ROOT'] . '/wp-config.php') ) {
+      require_once( $_SERVER['DOCUMENT_ROOT'] . '/wp-config.php' );
+    } else {
+      // trying probable locations of wp-config.php
+      if (file_exists( MYPATH . '/../../../wp-config.php' )) {
+        require_once( MYPATH . '/../../../wp-config.php' );
+      } elseif (file_exists( MYPATH . '/../../wp-config.php' )) {
+        require_once( MYPATH . '/../../wp-config.php' );
+      } elseif (file_exists( MYPATH . '/../wp-config.php' )) {
+        require_once( MYPATH . '/../wp-config.php' );
+      } elseif (file_exists( MYPATH . '/wp-config.php' )) {
+        require_once( MYPATH . '/wp-config.php' );
+      } else {
+        // shit, could not find wp-config.php
+        die();
+      }
+    }
+    if (file_exists($_SERVER['DOCUMENT_ROOT'] . '/wp-includes/wp-db.php') ) {
+      require_once( $_SERVER['DOCUMENT_ROOT'] . '/wp-includes/wp-db.php' );
+    } else {
+      // trying probable locations of wp-config.php
+      if (file_exists( MYPATH . '/../../../wp-includes/wp-db.php' )) {
+        require_once( MYPATH . '/../../../wp-includes/wp-db.php' );
+      } elseif (file_exists( MYPATH . '/../../wp-includes/wp-db.php' )) {
+        require_once( MYPATH . '/../../wp-includes/wp-db.php' );
+      } elseif (file_exists( MYPATH . '/../wp-includes/wp-db.php' )) {
+        require_once( MYPATH . '/../wp-includes/wp-db.php' );
+      } elseif (file_exists( MYPATH . '/wp-includes/wp-db.php' )) {
+        require_once( MYPATH . '/wp-includes/wp-db.php' );
+      } else {
+        // shit, could not find wp-includes/wp-db.php
+        die();
+      }
+    }
 
     global $wpdb;
     $table_name = $wpdb->prefix . 'cloudwok';
@@ -9,7 +42,7 @@
 
     if ($_SERVER['REQUEST_METHOD'] === 'POST') {
       // update
-     if(!array_diff(array('cwnonce','ref','code'), array_keys($_POST)) || ! wp_verify_nonce( $_POST['cwnonce'], $noncetxt ))
+     if(!array_diff(array('cwnonce','ref','code'), array_keys($_POST)) && wp_verify_nonce( $_POST['cwnonce'], $noncetxt ))
      {
           $ref=(int)$_POST['ref'];
           $code=$_POST['code'];
@@ -25,7 +58,7 @@
               array( '%d' )
           );
           echo "Updated";
-      } else if(!array_diff(array('cwnonce','code'), array_keys($_POST)) || ! wp_verify_nonce( $_POST['cwnonce'], $noncetxt )) {
+      } else if(!array_diff(array('cwnonce','code'), array_keys($_POST)) && wp_verify_nonce( $_POST['cwnonce'], $noncetxt )) {
          // create
         $wpdb->insert(
           $table_name,
@@ -69,7 +102,7 @@
      $_DELETE = array();
      parse_str(file_get_contents('php://input'), $_DELETE );
       // delete row
-     if(!array_diff(array('cwnonce','ref'), array_keys($_DELETE)) || ! wp_verify_nonce( $_DELETE['cwnonce'], $noncetxt ))
+     if(!array_diff(array('cwnonce','ref'), array_keys($_DELETE)) && wp_verify_nonce( $_DELETE['cwnonce'], $noncetxt ))
      {
          $ref=(int)$_DELETE['ref'];
          $wpdb->delete( $table_name, array( 'id' => $ref ) );
